+++ build.gradle
@@ -29,11 +29,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("seedu.sgsafe.SGSafe")
 }
 
 shadowJar {
-    archiveBaseName.set("duke")
+    archiveBaseName.set("sgsafe")
     archiveClassifier.set("")
 }
