+++ src/test/java/seedu/sgsafe/SGSafeTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.sgsafe;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class SGSafeTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
