+++ text-ui-test/EXPECTED.TXT
@@ -1,9 +1,3 @@
-Hello from
- ____        _        
-|  _ \ _   _| | _____ 
-| | | | | | | |/ / _ \
-| |_| | |_| |   <  __/
-|____/ \__,_|_|\_\___|
-
+This is SGSafe
 What is your name?
 Hello James Gosling
